% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_formula_2_adducts.R
\name{chem_formula_2_adducts}
\alias{chem_formula_2_adducts}
\title{Get adducts and isotopes for chemical formulas}
\usage{
chem_formula_2_adducts(
  chem.forms,
  isotopes,
  adducts,
  adduct.names = NULL,
  ion.mode = "positive",
  rel_to = 0,
  threshold = 20,
  verbose = TRUE
)
}
\arguments{
\item{chem.forms}{Vector of character strings with chemical formulas.}

\item{isotopes}{dataframe with stable isotopes (see
\code{\link[enviPat]{isotopes}})}

\item{adducts}{dataframe with common adducts (see
\code{\link[enviPat]{adducts}}).}

\item{adduct.names}{Character vector with adducts to be calculated
(see \code{\link[enviPat]{adducts}} for common adducts). If set to NULL,
ion.mode needs to be specified.}

\item{ion.mode}{Character of ionisation mode ("positive", "negative"). If
positive adduct.names is set to c("M+H", "M+Na", "M+K"), if negative
adduct.names is set to c("M-H", "M+Cl").}

\item{rel_to}{Integer (0, 1, 2, 3 or 4). Probability definition
(see \code{\link[enviPat]{isopattern}}).}

\item{threshold}{Numeric (1-100). Probability below which isotope peaks can
be omitted (see \code{\link[enviPat]{isopattern}}).}

\item{verbose}{Logical. Print out additional information.}
}
\value{
data frame with the following information about each feature:
\itemize{
  \item chemical.form.adduct: chemical formula of adduct
  \item chemical.form: original chemical formula
  \item adduct.name: adduct used
  \item charge: charge of adduct
  \item chemical.form.isotope: chemical formula of isotope of adduct
  \item m.z: theoretical mz value of isotope
  \item abundance: theoretical abundance of isotope peak. Most abundant
  isotope has value of 100
  \item id: unique identifier (<chemical.form>.<chemical.form.isotope>)
}
}
\description{
Calculates chemical formulas, mz values and theoretical abundances of
isotope peaks of adducts of chemical formulas. Based on functions in R
package enviPat.
}
\examples{
library(enviPat)
data("isotopes")
data("adducts")

chem_formula_2_adducts(chem.forms = "C6H12O6",
                       adduct.names = c("M+H", "M+Na", "M+2H"),
                       adducts = adducts,
                       isotopes = isotopes,
                       threshold = 5)
}
