% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_feature_intensities.R
\name{combine_feature_intensities}
\alias{combine_feature_intensities}
\title{Combine feature intensities of several spectra}
\usage{
combine_feature_intensities(feature.files, id.samples = NULL,
  pheno = NULL, cor.cutoff = 0.9, verbose = FALSE)
}
\arguments{
\item{feature.files}{Character vector. Names of files with feature
intensities as generated by \code{\link{extract_feature_intensity}}.}

\item{id.samples}{Character vector. Identifiers of each file. If NULL, ids
will be identified as basenames of files.}

\item{pheno}{data.frame. Information about each sample. Rownames need to
correspond to id.samples. If NULL, pheno will only contain ids.}

\item{cor.cutoff}{Numeric. Cutoff of the pair-wise absolute correlation
(default: 0.9).}

\item{verbose}{Logical. Print out additional information.}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object.
}
\description{
Load extracted information about features such as intensities, peak
detection and measured mz values and combine them with phenotype information
into a \code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object. Features are removed if they have missing mz values or intensities
for all samples or constant intensity across all samples.
}
