% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{compute_post}
\alias{compute_post}
\title{internal function
copy of function IPA:::compute.post}
\usage{
compute_post(allsampcomp, no.its, burn, Nc)
}
\description{
internal function
copy of function IPA:::compute.post
}
\keyword{internal}
