% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{compute_posterior_prob}
\alias{compute_posterior_prob}
\title{Compute posterior probabilities for putative theoretical features}
\usage{
compute_posterior_prob(prior.prob, dat, add.m = NULL, iso.m = NULL,
  bio.m = NULL, delta.add = 0.5, delta.iso = 0.5, delta.bio = 1,
  ratio.tol = 0.8, no.its = 1100, burn = 100)
}
\arguments{
\item{prior.prob}{matrix with prior probabilities for each measured feature
in rows and theoretical features in columns (e.g. as generated by the
function \code{\link{compute_prior_prob}}).}

\item{dat}{data.frame with information about measured features with columns
id (= unique identifier), mz (= measured mz value), intensity (= measured
intensity).}

\item{add.m}{matrix with information about adduct connections (e.g. as
generated by function \code{\link{generate_connectivity_matrix}}).)}

\item{iso.m}{matrix with information about isotope connections (e.g. as
generated by function \code{\link{generate_connectivity_matrix}}).)}

\item{bio.m}{matrix with information about biochemical connections (e.g. as
generated by function \code{\link{generate_connectivity_matrix}}).)}

\item{delta.add}{Numeric. Confidence on the information encoded in add.m
(smaller value means higher confidence).}

\item{delta.iso}{Numeric. Confidence on the information encoded in iso.m
(smaller value means higher confidence).}

\item{delta.bio}{Numeric. Confidence on the information encoded in bio.m
(smaller value means higher confidence).}

\item{ratio.tol}{Numeric. Minimum accepted ratio between thereoretical and
observed intensity ratios between isotopes (default = 0.8).}

\item{no.its}{Numeric. Number of iterations to be performed by the Gibbs
sampler (default = 1100).}

\item{burn}{Numeric. Number of initial iterations to be ignored when
computing the posterior probabilities (default = 100).}
}
\value{
matrix with measured features in rows and theoretical possible features in
columns. The cell i,j contains the posterior probability that measured
feature i belongs to theoretical feature j.
}
\description{
Estimates posterior probabilities of all possible theoretical features for each
measured feature based on prior probabilities and additional information about
adduct and isotope connections using a Gibbs sampler. This function is based
on code of the internal function IPA.sampler.Add.Iso.Bio in the R package
IPA (available on github at \url{https://github.com/francescodc87/IPA}).
}
