% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{compute_prior_prob}
\alias{compute_prior_prob}
\title{Compute prior probabilities for putative theoretical features}
\usage{
compute_prior_prob(
  hits.m,
  info.features,
  dat,
  pk = NULL,
  ppm,
  ppm.unknown = NULL,
  pr.limit = 1e-05
)
}
\arguments{
\item{hits.m}{matrix with information about overlapping theoretical features
(e.g. as generated by the function \code{\link{find_hits}}).}

\item{info.features}{data.frame with information about theoretical possible
features (e.g. as generated by the function
\code{\link{chem_formula_2_adducts}}).}

\item{dat}{data.frame with information about measured features with columns
id (= unique identifier), mz (= measured mz value), intensity (= measured
intensity).}

\item{pk}{Numeric vector with probabilities of initial confidence of the
presence of each theoretical feature.}

\item{ppm}{Numeric. Resolution of mass spectrometer.}

\item{ppm.unknown}{Numeric. Resolution assigned to unknown theoretical
features (i.e. not contained in info.features).}

\item{pr.limit}{Numeric. Lowest probability not set to 0 (default = 1e-05).}
}
\value{
matrix with measured features in rows and theoretical possible features in
columns. The cell i,j contains the prior probability that measured feature i
belongs to theoretical feature j.
}
\description{
Estimates prior probabilities of all possible theoretical features for each
measured feature.
}
\examples{
data("se.example")
data("info.features")

dat = prepare_data_for_annotation(se = se.example)
hits.m = find_hits(info.features = info.features,
                   dat = dat,
                   ppm = 20)

prior.prob = compute_prior_prob(hits.m = hits.m,
                                info.features = info.features,
                                dat = dat,
                                ppm = 20)
}
