% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_chem_formula.R
\name{correct_chem_formula}
\alias{correct_chem_formula}
\title{Correct chemical formulas}
\usage{
correct_chem_formula(chem.forms, isotopes)
}
\arguments{
\item{chem.forms}{Vector of character strings with chemical formulas.}

\item{isotopes}{dataframe with stable isotopes (see
\code{\link[enviPat]{isotopes}})}
}
\value{
Vector of character strings with corrected chemical formulas
}
\description{
Orders elements in chemical formulas alphabetically and adds 1 for elements
that occur only once.
}
\examples{
library(enviPat)
data("isotopes")

correct_chem_formula(chem.forms = c("C44H92NO6P", "C6H12O6"),
                     isotopes = isotopes)
}
