% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature_intensity.R
\name{extract_feature_intensity}
\alias{extract_feature_intensity}
\title{Extract intensities of specific features from a single spectrum}
\usage{
extract_feature_intensity(spectrum.file, scanrange = c(1, 1),
  info.features, ppm = 10, num.pts = 5, oneside.min = 1,
  verbose = TRUE, res.dir = NULL)
}
\arguments{
\item{spectrum.file}{Character. Name of mzML file to read.}

\item{scanrange}{Numeric vector of length 2. Scan range to read
(see \code{\link[xcms]{xcmsRaw}}).}

\item{info.features}{data.frame with information about features (e.g. as
generated by the function \code{\link{chem_formula_2_adducts}}).}

\item{ppm}{Numeric. Resolution of mass spectrometer.}

\item{num.pts}{Numeric. Minimum number of points needed to search for a peak
(see \code{\link[FTICRMS]{locate.peaks}}).}

\item{oneside.min}{Numeric. Minimum number of points needed on each side of
the local maximum (see \code{\link[FTICRMS]{locate.peaks}}).}

\item{verbose}{Logical. Print out additional information.}

\item{res.dir}{Character. Name of directory where result is saved as RDS
object (using the basename of the spectrum file). If NULL, a data.frame with
the results will be returned.}
}
\value{
if res.dir = NULL, data frame with information for each feature:
\itemize{
  \item id: feature identifier (corresponds to chemical.form.isotope in
  info.features)
  \item mz.th: theoretical mz value
  \item intensity: extracted intensity
  \item mz.meas: mz value at center of peak (or closest measured mz value)
  \item peak.detected: 0 = no peak detected, 1 = peak detected, NA = no or
  not enough mz values available for feature
  }
}
\description{
For each of the given features local maxima will be identified in a region
of 2ppm around the theoretical mz value using the algorithm implemented in
the function \code{\link[FTICRMS]{locate.peaks}}. The intensity value at the
local maximum with maximal intenstity within 1ppm will be extracted.
If no peak could be detected, the intensity at the closest measured mz value
is used.
}
\examples{
data("info.features")
res.dir = tempdir()
mzml.file = system.file("extdata",
                        "20121015b__002__07__ME.mzML",
                        package = "preprocessHighResMS")

extract_feature_intensity(spectrum.file = mzml.file,
                          scanrange = c(1, 2),
                          info.features = info.features,
                          ppm = 20,
                          res.dir = res.dir)
}
