% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{generate_connectivity_matrix}
\alias{generate_connectivity_matrix}
\title{Generate connectivity matrices}
\usage{
generate_connectivity_matrix(info.features, type, ids.use = NULL,
  ratios = FALSE)
}
\arguments{
\item{info.features}{data.frame with information about theoretical possible
features (e.g. as generated by the function
\code{\link{chem_formula_2_adducts}}).}

\item{type}{Character. Type of connectivity matrix to generate ("adducts",
"isotopes").}

\item{ids.use}{Character vector. Feature identifiers to be used for
connectivity matrix. If NULL connectivity matrix is build using all features
in info.features.}

\item{ratios}{Logical. Should the isotope connectivity matrix contain the
expected abundance ratios? (deafult = TRUE).}
}
\value{
matrix with theoretical possible features in rows and columns.
In the adduct matrix, the cell i,j contains 1, if features i and j are
monoisotopic isotopes and adducts of the same chemical formula and 0
otherwise.
In the isotope matrix, the cell i,j contains 1, if features i and j are
isotopes of the same adduct (or the ratio of the theoretical abundances if
ratios = TRUE) and 0 otherwise.
}
\description{
Generates matrices summarizing connections between adducts and isotopes.
}
\examples{
data("info.features")

add.m = generate_connectivity_matrix(info.features = info.features,
                                     type = "adducts")

iso.m = generate_connectivity_matrix(info.features = info.features,
                                     type = "isotopes")
}
