% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_formula_2_adducts.R
\name{get_info_isotopes}
\alias{get_info_isotopes}
\title{Isotope information}
\usage{
get_info_isotopes(chem.forms, isotopes, rel_to = 0, threshold = 20)
}
\arguments{
\item{chem.forms}{Vector of character strings with chemical formulas.}

\item{isotopes}{dataframe with stable isotopes (see
\code{\link[enviPat]{isotopes}}).}

\item{rel_to}{Integer (0, 1, 2, 3 or 4). Probability definition
(see \code{\link[enviPat]{isopattern}}).}

\item{threshold}{Numeric (1-100). Probability below which isotope peaks can be omitted
(see \code{\link[enviPat]{isopattern}}).}
}
\value{
data frame with the following information about each feature:
\itemize{
  \item chemical.form.adduct: chemical formula of adduct
  \item chemical.form: original chemical formula
  \item adduct.name: adduct used
  \item chemical.form.isotope: chemical formula of isotope of adduct
  \item m.z: theoretical mz value of isotope
  \item abundance: theoretical abundance of isotope peak. Most abundant
  isotope has value of 100
  \item id: unique identifier (<chemical.form>.<chemical.form.isotope>)
}
}
\description{
Based on functions in R package enviPat.
}
\keyword{internal}
