% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info.features}
\alias{info.features}
\title{Feature information}
\format{data.frame as generated by the function
\code{\link{chem_formula_2_adducts}}}
\source{
Eiden et al. 2015 Genome Medicine 7:63
}
\usage{
info.features
}
\description{
Information about features based on three internal standards (C40H81N2O6P1,
C44H92N1O6P1 and C45H94N1O6P1) that have mz values between 700 and 800.
}
\keyword{datasets}
