% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature_intensity.R
\name{load_spectrum}
\alias{load_spectrum}
\title{Load spectrum}
\usage{
load_spectrum(spectrum.file, scanrange = c(1, 1))
}
\arguments{
\item{spectrum.file}{Character. Name of mzML file to read.}

\item{scanrange}{Numeric vector of length 2. Scan range to read
(see \code{\link[xcms]{xcmsRaw}}).}
}
\value{
data frame with spectrum information:
\itemize{
  \item mz: mz value
  \item intensity: intensity
}
}
\description{
Based on functions in Bioconductor package xcms.
}
\keyword{internal}
