% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{multsample}
\alias{multsample}
\title{internal function
copy of function IPA:::multsample}
\usage{
multsample(P)
}
\description{
internal function
copy of function IPA:::multsample
}
\keyword{internal}
