% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{prepare_data_for_annotation}
\alias{prepare_data_for_annotation}
\title{Prepare input for annotation}
\usage{
prepare_data_for_annotation(se)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}
}
\value{
data.frame with the following information about measured features:
\itemize{
\item id: unique identifier
\item mz: measured mz value
\item intensity: measured intensity
\item no.peak.detected: number of detected peaks.
}
}
\description{
Extract information for annotation from SummarizedExperiment object. For
each feature, the median mz value across all samples will be used as
measured mz value and the sum of all intensities for the intensity value.
}
\examples{
data("se.example")
dat = prepare_data_for_annotation(se = se.example)
}
