% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_features.R
\name{remove_features}
\alias{remove_features}
\title{Remove features based on different criteria}
\usage{
remove_features(se, assay, method, freq = 0.25, verbose = FALSE)
}
\arguments{
\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}

\item{assay}{Character or integer. Name or number of assay to be used for
filtering.}

\item{method}{Method to determine features to be removed: "constant",
"missing", "zero", "identical.peaks".}

\item{freq}{Numeric. If more or equal than freq*100 % of the samples fulfill
criterion the feature is removed (not used if method = "constant").}

\item{verbose}{Logical. Should number of removed features be reported?}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with features removed
}
\description{
Remove features with no variability (method = "constant"), missing values
(method = "missing"), low intensity (method = "zero") or identical peaks
(method = "identical.peaks").
}
\examples{
data("se.example")
se.example = remove_features(se = se.example,
                             assay = "mz",
                             method = "identical.peaks")
}
