% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se.example}
\alias{se.example}
\title{Feature level metabolomics data}
\format{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object}
\source{
\url{https://www.omicsdi.org/dataset/metabolights_dataset/MTBLS162}
}
\usage{
se.example
}
\description{
Metabolomics data on the feature level from two quality control samples of
the study MTBLS162 in the MetaboLights database reduced to mz values between
700 and 800.
}
\keyword{datasets}
