% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{assign_features}
\alias{assign_features}
\title{Assign measured feature to theoretical features}
\usage{
assign_features(post.prob, dat, cutoff.prob = 0.8)
}
\arguments{
\item{post.prob}{matrix with prior probabilities for each measured feature
in rows and theoretical features in columns (e.g. as generated by the
function \code{\link{compute_posterior_prob}}).}

\item{dat}{data.frame with information about measured features with columns
id (= unique identifier), mz (= measured mz value), intensity (= measured
intensity).}

\item{cutoff.prob}{Numeric. Probability cutoff which needs to be reached to
keep the assignment of measured and theoretical feature (default = 0.8).}
}
\value{
data.frame with the following information about each assigned feature:
\itemize{
\item f.meas: identifier of measured feature
\item f.theo: identifier of assigned theoretical feature
\item prob: posterior probability
}
}
\description{
Based on estimated posterior probabilities each measured feature is assigned
to the most probable theoretical feature. Measured features are removed if
the maximum probability is below the specified cutoff. In rare cases,
several measured features are assigned to the same theoretical feature and
the measured feature with the largest number of detected peaks or highest
intensity will be kept.
}
\examples{
data("se.example")
data("info.features")

dat = prepare_data_for_annotation(se = se.example)
hits.m = find_hits(info.features = info.features,
                   dat = dat,
                   ppm = 20)

prior.prob = compute_prior_prob(hits.m = hits.m,
                                info.features = info.features,
                                dat = dat,
                                ppm = 20)
add.m = generate_connectivity_matrix(info.features = info.features,
                                     type = "adducts")

set.seed(20200402)
post.prob = compute_posterior_prob(prior.prob = prior.prob,
                                   dat = dat,
                                   add.m = add.m,
                                   delta.add = 0.1)

info.assigned.use = assign_features(post.prob = post.prob,
                                    dat = dat)
}
