% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{summarize_metabolites}
\alias{summarize_metabolites}
\title{Summarize feature based data into metabolite based data}
\usage{
summarize_metabolites(info.assigned, info.features, se)
}
\arguments{
\item{info.assigned}{data.frame with information about assignment of measured
and theoretical features (e.g. as generated by the function
\code{\link{assign_features}}).}

\item{info.features}{data.frame with information about theoretical possible
features (e.g. as generated by the function
\code{\link{chem_formula_2_adducts}}).}

\item{se}{\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with feature based metabolomics data.}
}
\value{
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
object with metabolite based metabolomics data.
}
\description{
Combine the feature based metabolomics data to metabolite data by summing up
the intensitites of all features assigned to a particular metabolite.
}
